+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added the exposure sliders to the planar reflection probe preview
 - Added a warning and workaround instructions that appear when you enable XR single-pass after the first frame with the XR SDK.
 - Added an "enable" toggle to the SSR volume component.
+- Added support of cookie baking and add support on Disc light.
 
 ### Fixed
 - Fix Changelog